requirements.txt
requests-html
python-telegram-bot==13.11
chromedriver-autoinstaller
selenium